# open-api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080/save-species","description":"Generated server url"}],"paths":{"/v1/species":{"get":{"tags":["species-controller"],"summary":"Get all species available. Method available for any authenticated user. Results are cached for 10 minutes","operationId":"getAllSpecies","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SpeciesListDto"}}}}}},"put":{"tags":["species-controller"],"summary":"Create given species. Required ADMIN role","operationId":"getSpecies","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SpeciesDto"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/v1/donations":{"get":{"tags":["donations-controller"],"summary":"Get list of all donations for authenticated user","operationId":"getDonationsForUser","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DonationDetailsListDto"}}}}}},"post":{"tags":["donations-controller"],"summary":"Add information about new donation for authenticated user","operationId":"recordDonation","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DonationDto"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/v1/species/{speciesName}":{"get":{"tags":["species-controller"],"summary":"Get given species. Required ADMIN role","operationId":"getSpecies_1","parameters":[{"name":"speciesName","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SpeciesDto"}}}}}},"delete":{"tags":["species-controller"],"summary":"Delete given species. Required ADMIN role","operationId":"deleteSpecies","parameters":[{"name":"speciesName","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/v1/donations/{donationId}":{"delete":{"tags":["donations-controller"],"summary":"Delete donation by donationId for authenticated user","operationId":"withdrawDonation","parameters":[{"name":"donationId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"SpeciesDto":{"type":"object","properties":{"name":{"type":"string"},"categories":{"type":"array","items":{"type":"string"}},"imagePath":{"type":"string"}}},"DonationDto":{"type":"object","properties":{"species":{"type":"string"},"amount":{"type":"string"}}},"SpeciesListDto":{"type":"object","properties":{"speciesDtos":{"type":"array","items":{"$ref":"#/components/schemas/SpeciesDto"}}}},"DonationDetailsDto":{"type":"object","properties":{"donationId":{"type":"string"},"userId":{"type":"string"},"species":{"type":"string"},"amount":{"type":"string"}}},"DonationDetailsListDto":{"type":"object","properties":{"donationDetailsDtoList":{"type":"array","items":{"$ref":"#/components/schemas/DonationDetailsDto"}}}}}}}